Started to use doubles for layout and painting
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@661 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_abstract_scale_draw.h
@@ -102,7 +102,7 @@ protected:
 
        \sa drawBackbone(), drawLabel()
     */  
-    virtual void drawTick(QPainter *painter, double value, int len) const = 0;
+    virtual void drawTick(QPainter *painter, double value, double len) const = 0;
 
     /*!
       Draws the baseline of the scale
